+++ lib/extensions/navigatgor_key.dart
@@ -10,13 +10,24 @@ extension RouterDelegateExtensions on GlobalKey<NavigatorState> {
   }) =>
       _delegate.popRoute();
 
-  void navigateTo(
+  FutureOr<void> navigateTo(
     ModularPage page, {
     bool clearHistory = false,
     bool removeCurrent = false,
   }) =>
       _delegate.navigateTo(
         page: page,
+        clearHistory: clearHistory,
+        removeCurrent: removeCurrent,
+      );
+
+  FutureOr<void> navigateByUri({
+    required Uri uri,
+    bool clearHistory = false,
+    bool removeCurrent = false,
+  }) =>
+      _delegate.navigateToUri(
+        uri: uri,
         clearHistory: clearHistory,
         removeCurrent: removeCurrent,
       );