+++ lib/modules/root_module.dart
@@ -5,7 +5,11 @@ abstract class RootModule extends BaseModule {
   ModularRoute get homeRoute;
   ModularRoute get notFoundRoute;
 
-  const RootModule();
+  const RootModule({
+    FutureOr<bool> Function(
+            ModularHistory route, ModularRouterDelegate delegate)?
+        guard,
+  }) : super(guard);
 
   @override
   String get route => "/";