+++ lib/modules/sub_module.dart
@@ -1,8 +1,13 @@
 part of modular_navigation;
 
 abstract class SubModule extends BaseModule {
-  final String parentRoute;
-  const SubModule(this.parentRoute);
+  final BaseModule parentModule;
+  const SubModule({
+    required this.parentModule,
+    FutureOr<bool> Function(
+            ModularHistory route, ModularRouterDelegate delegate)?
+        guard,
+  }) : super(guard);
 
   @override
   Iterable<BaseModule> get subModules => [];