+++ lib/navigator/modular_history.dart
@@ -5,4 +5,7 @@ class ModularHistory<TPageParameters extends PageParameters,
   final ModularRoute route;
   final TModularPage page;
   const ModularHistory(this.route, this.page);
+
+  Uri get uri =>
+      Uri.parse(route.route).replace(queryParameters: page.parameters.map);
 }